+++ src/com/sam/TubeCirc.java
@@ -3,7 +3,7 @@ package com.sam;
 public class TubeCirc extends BarCirc {
     int wallThic;
 
-    public TubeCirc(int matType, int diameter, double length, int wallThic) {
+    public TubeCirc(int matType, int diameter, int length, int wallThic) {
         super(matType, diameter, length);
         this.setMaterial(matType);
         this.diameter = diameter;
