+++ src/com/sam/TubeRec.java
@@ -2,7 +2,7 @@ package com.sam;
 
 public class TubeRec extends BarRec{
     int wallThic;
-    public TubeRec(int matType, int sideA, int sideB, double length, int wallThic) {
+    public TubeRec(int matType, int sideA, int sideB, int length, int wallThic) {
         super(matType, sideA, sideB, length);
 
         this.setMaterial(matType);
